migrated to glutin 0.26, fixed capsule visualization a bit

+++ rg3d-ui/src/message.rs
@@ -1028,7 +1028,7 @@ pub enum MouseButton {
     Left,
     Right,
     Middle,
-    Other(u8),
+    Other(u16),
 }
 
 pub enum OsEvent {
